# merkle: fix proof performance