The target need only be hashed once.

Before: BenchmarkProof-10 1 53486735083 ns/op (53.486735083s)
After:  BenchmarkProof-10 4 299320958 ns/op (0.299320958s)